% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\docType{data}
\name{flpalette}
\alias{flpalette}
\alias{flpalette_colours}
\alias{flpalette_grads}
\title{A highly-contrasted palette to use in ggplotFL}
\format{An object of class \code{character} of length 7.}
\usage{
flpalette

flpalette_colours(n = length(flpalette))

flpalette_grads(palette = flpalette, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Palette subset to create a gradual scale from, defaults to *flpalette*.}

\item{reverse}{Should the palette be reversed, FALSE.}

\item{...}{Elements to subset from palette, by name or position.}
}
\value{
A named vector of colors and HEX codes, or a function to obtain a gradient of colors fo a given length.
}
\description{
Plot methods defined in the ggplotFL package make use by default of a palette
with a high contrast, useful to separate time series or categories. The palette
consist of seven colours: red, blue, green, violet, orange, yellow and brown.
This palette can be inspected at the [colorbrewe2.org site](http://colorbrewer2.org/?type=qualitative&scheme=Set1&n=7#type=qualitative&scheme=Set1&n=7).
}
\details{
The palette is accessible as a named vector, *flpalette*, with the corresponding
HEX codes. Two functions are also available to manipualte the palette. One to
extract a subset of the palette, *flpalette_colours*, and another to create a
gradation of colours between two or more of the palette colours, *flpalette_grads*.
}
\examples{
# CHECK flpalette
 flpalette
flpalette_colours()
flpalette_colours(5)
flpalette_colours(2:3)
flpalette_grads(flpalette_colours(3))(20)
}
\seealso{
\link{FLComp}
}
\author{
The FLR Team
}
\keyword{classes}
