% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{human_numbers}
\alias{human_numbers}
\title{A ggplot2 number formatter}
\usage{
human_numbers(x = NULL, smbl = "", signif = 1)
}
\arguments{
\item{x}{An object of class \code{numeric}}

\item{smbl}{A character or symbol to be added prior to the number, e.g. an euro sign.}

\item{signif}{Number of significant figures}
}
\value{
A formatted character vector
}
\description{
This function formats numbers for output in a 'human' way.
}
\seealso{
\link[ggplot2]{labeller}
}
\author{
Iago Mosqueira (EC JRC)
}
\keyword{dplot}
