% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Shows integer values in a programmatic and scalable fashion.}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{Number of breaks}
}
\value{
A function to be called
}
\description{
This function provides sensible breaks for integers
}
\seealso{
\link[ggplot2]{labeller}
}
\author{
Iago Mosqueira (WMR)
}
\keyword{dplot}
