% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{label_flqs}
\alias{label_flqs}
\title{A ggplot2 labeller for FLQuants}
\usage{
label_flqs(x, ...)
}
\arguments{
\item{x}{An object of class \code{FLQuants}}

\item{drop}{Character string to be dropped from the label shen found in the
*units* slot. Defaults to c("NA", "NC", "m", "z", "prop")}
}
\value{
A ggplot2 labeller function
}
\description{
Plots of \code{\link{FLQuants}} objects use by default \code{\link{facet_grid}}
to separate the different elements. This function generates facet labels that
have both the element name and the units of measurement of each, as stored
in the 'units' slot of each \code{\link{FLQuant}}.
}
\details{
Certain *units* are dropped from the label, as being uninformative: "NA", "NC",
"m", "z", and "prop". This can be selected with the *drop* argument.
}
\seealso{
\link[ggplot2]{labeller}
}
\author{
Iago Mosqueira (EC JRC)
}
\keyword{dplot}
