% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{eqlabel}
\alias{eqlabel}
\alias{modlabel}
\title{Functions create labels for FLSR models}
\usage{
eqlabel(model, param)

modlabel(model, param)
}
\arguments{
\item{model}{a list of model formulas}

\item{param}{a list of FLPar objects for the params slot}
}
\description{
These functions create labels from FLSR models and params to be used, for
example, on the legend of the ggplot-based plot method for FLSRs
}
\examples{
 data(nsher)
 srs <- FLSRs(sapply(c('ricker', 'bevholt'), function(x) {
   y <- nsher
   model(y) <- x
   return(fmle(y))
 }))
 eqlabel(model=lapply(srs, model),
   param=lapply(srs, params))


 modlabel(model=lapply(srs, model),
   param=lapply(srs, params))

}

