% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{plotKobe}
\alias{plotKobe}
\title{Kobe plot of biomass and fishing mortality over MSY}
\usage{
plotKobe(x, refpts, X = SB/SB[MSY] ~ ssb/SBMSY, Y = F/F[MSY] ~ fbar/FMSY)
}
\arguments{
\item{x}{An object of class FLStock}

\item{refpts}{Reference points, class FLPar or list.}

\item{X}{Formula for x axis metric, defaults to SB/SB[MSY] ~ ssb/SBMSY.}

\item{Y}{Formula for y axis metric, defaults to F/F[MSY] ~ fbar/FMSY.}
}
\value{
A ggplot object
}
\description{
'Kobe' plot is the common name given to a phase plot that show the time series and
current status of stock status using two metrics: one for stock status on the 
x-axis, generally SB/SBMSY, and one on exploitation level on the y-axis, F/FMSY.
The plot is divided in four quadrants, each representing a different combination
of stock status and exploitation level. The quadrants are usually coloured green 
(healthy stock and exploitation), yellow (healthy stock, high exploitation), orange 
(overfished stock, healthy exploitation) and red (overfished stock, high 
exploitation).
}
\details{
The function allows the user to define the metrics to be used on each axis, by using 
a formula interface. For each axis, a formula specifies the name of the metric and 
its computation, using available slots and metrics for the FLStock object, as well 
as reference points provided in the `refpts` argument. See examples below on the 
usage of the formula interface.
}
\examples{
data(ple4)
# Plot with default MSY refpts
plotKobe(ple4, refpts=FLPar(FMSY=0.28, SBMSY=555000))
# Plot with custom metric on x axis
plotKobe(ple4, refpts=FLPar(FMSY=0.28, SB0=8e5), X=SB/SB[0]~ssb/SB0)
}
\seealso{
[FLStock] [FLPar]
}
\author{
Iago Mosqueira (WMR)
}
\keyword{hplot}
