% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{plotRunstest}
\alias{plotRunstest}
\alias{plotRunstest,FLQuants,missing-method}
\alias{plotRunstest,FLQuants,FLQuants-method}
\alias{plotRunstest,FLQuant,FLQuant-method}
\title{Plot the runs test result for one or more time series}
\usage{
plotRunstest(fit, obs, ...)

\S4method{plotRunstest}{FLQuants,missing}(fit, combine = TRUE)

\S4method{plotRunstest}{FLQuants,FLQuants}(fit, obs, combine = TRUE)

\S4method{plotRunstest}{FLQuant,FLQuant}(fit, obs, combine = TRUE)
}
\arguments{
\item{fit}{The result of a model fit.}

\item{obs}{The observations used in the fit.}

\item{...}{Extra arguments.}

\item{combine}{Should ages be combined by addition, defaults to TRUE.}
}
\value{
An object of class ggplot2::gg
}
\description{
Plot the runs test result for one or more time series
}
\examples{
data(nsher)
plotRunstest(fitted(nsher), rec(nsher))
}
