% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{plotXval}
\alias{plotXval}
\title{Plot of FLIndices cross-validation by retrospective hindcast}
\usage{
plotXval(x, y = "missing", order = "inverse")
}
\arguments{
\item{x}{An *FLIndices* object of the original observations.}

\item{y}{A list contaning *FLIndices* objects returned by *a4ahcxval*.}

\item{order}{Order in which retrospective runs are stored, defaults to"inverse".}
}
\value{
A ggplot object
}
\description{
Plot of FLIndices cross-validation by retrospective hindcast
}
\examples{
# SEE vignette
}
