% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{pubpng}
\alias{pubpng}
\title{Output ggplot object to PNG files with good quality settings.}
\usage{
pubpng(file, plot, width = 1600, height = 1400, res = 200)
}
\arguments{
\item{file}{Output file path and name.}

\item{plot}{ggplot object or command.}

\item{width}{Width of plot, in pixels. Defaults to 1600.}

\item{height}{Height of plot, in pixels. Defaults to 1400.}

\item{res}{Resolution, in ppi. Deagults to 200.}
}
\value{
TRUE if succesful, while file is saved to disk.
}
\description{
Output ggplot object to PNG files with good quality settings.
}
\seealso{
grDevices::png
}
