% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{scale_fill_flr}
\alias{scale_fill_flr}
\title{High contrast discrete and continuous palettes}
\usage{
scale_fill_flr(palette = flpalette, discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Palette subset to create a gradual scale from, defaults to *flpalette*.}

\item{discrete}{Is the palette to be applied to a discrete variable, TRUE.}

\item{reverse}{Should the palette be reversed, FALSE.}

\item{...}{Other arguments to be passed to the scale functions}
}
\value{
A function.
}
\description{
Discrete and continuous versions of the standard ggplotFL palette.
See \link{flpalette} fo further details.
}
\seealso{
\link{flpalette}
}
\author{
The FLR Team
}
\keyword{color}
