% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/special.R
\name{cohcorrplot}
\alias{cohcorrplot}
\alias{cohcorrplot,FLQuant-method}
\alias{cohcorrplot,FLCohort-method}
\title{cohcorrplot}
\usage{
cohcorrplot(x, ...)

\S4method{cohcorrplot}{FLQuant}(x)

\S4method{cohcorrplot}{FLCohort}(x, diag_size = 16, lower_size = 6)
}
\arguments{
\item{x}{An object with the abundance at age information. FLQuant or FLCohort.}

\item{...}{Any extra arguments}

\item{diag_size}{Font size for labels in diagonal row}

\item{lower_size}{Font size for labels in lower triangle}
}
\description{
A correlation plot that show and quantifies correlation along
cohorts. Typically used on catch or survey abundances-at-age.
}
\details{
The method prints a plot assembled as a combination of grid elements, but
reurns it as a \emph{gg} object.
}
\examples{
data(ple4)
cohcorrplot(catch.n(ple4))
cohcorrplot(FLCohort(stock.n(ple4)))
}
\author{
The FLR Team
}
\keyword{methods}
