% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_flpar}
\alias{geom_flpar}
\title{Horizontal lines for FLPar objects}
\usage{
geom_flpar(mapping = NULL, data = NULL, ..., x, na.rm = FALSE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
This `geom` shows a horizontal line for each `param` in a `FLPar` object and
labels it using the `param` dimnames
}
\section{Aesthetics}{

`geom_flpar` understands the following aesthetics (required aesthetics are in bold):
- `*x*`
- `label`
- `alpha`
- `colour`
- `linetype`
- `size`
}

\examples{
data(ple4)
plot(ssb(ple4)) + geom_flpar(data=FLPar(Blim=300000), aes(x=1960))
plot(ssb(ple4)) + geom_flpar(data=FLPar(Blim=300000), x=1960)
plot(ssb(ple4)) + geom_flpar(data=FLPar(Blim=300000), x=2015)
plot(ssb(ple4)) + geom_flpar(data=FLPar(Blim=300000, Bpa=230000), aes(x=1960))
}
