% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_worm}
\alias{geom_worm}
\title{A geom for adding worms to probability intervals from geom_flquantiles}
\usage{
geom_worm(
  data,
  mapping = aes(colour = iter),
  ...,
  stat = "identity",
  position = "identity",
  na.rm = FALSE
)
}
\arguments{
\item{data}{Subset of data, select from full object using iter().}

\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}} or
\code{\link[ggplot2:aes_]{aes_()}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}
}
\description{
A geom for adding worms to probability intervals from geom_flquantiles
}
\section{Aesthetics}{

`geom_worm` understands the following aesthetics (required aesthetics are in bold):
- `colour`
- `linetype`
- `size`
}

\examples{
data(ple4)
x <- rlnorm(200, log(catch(ple4)), 0.3)
plot(x) + geom_worm(data=iter(x, 1:4))

x <- FLQuants(C=rlnorm(200, log(catch(ple4)), 0.3),
  F=rlnorm(200, fbar(ple4), 0.2))
plot(x) + geom_worm(data=iter(x, 1:4))
}
