% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot,FLQuant,missing-method}
\alias{plot,FLQuant,missing-method}
\alias{plot,FLQuants,missing-method}
\alias{plot,FLSR,missing-method}
\alias{plot,FLStock,FLPar-method}
\alias{plot,FLStock,missing-method}
\alias{plot,FLStocks,missing-method}
\title{ggplot versions of FLR class plot() methods}
\usage{
\S4method{plot}{FLQuant,missing}(x, main = "", xlab = "", ylab = "",
  na.rm = FALSE, probs = c(0.1, 0.25, 0.5, 0.75, 0.9), type = 7)

\S4method{plot}{FLQuants,missing}(x, main = "", xlab = "", ylab = "",
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9), na.rm = TRUE, type = 7)

\S4method{plot}{FLStock,missing}(x, main = "", xlab = "", ylab = "", ...)

\S4method{plot}{FLStocks,missing}(x, main = "", xlab = "", ylab = "", ...)

\S4method{plot}{FLStock,FLPar}(x, y, ...)

\S4method{plot}{FLSR,missing}(x, y, ...)
}
\description{
New basic plot methods for some FLR classes are defined in ggplotFL.
}
\examples{
# Plot a single FLQuant
 data(ple4)
 plot(catch.n(ple4))

 # Plot an FLQuant with iters, shows quantiles
 flq <- rnorm(100, catch(ple4), 60000)
 plot(flq)

 # Specify quantiles, default is c(0.10, 0.25, 0.50, 0.75, 0.90)
 plot(flq, probs=c(0.05, 0.40, 0.50, 0.60, 0.95))
# Plot an FLQuants created from ple4 FLStock
 data(ple4)
 plot(FLQuants(SSB=ssb(ple4), rec=rec(ple4)))
# plot of an FLStock
 data(ple4)
 plot(ple4)
# plot for FLStocks
 data(ple4)
 pls <- FLStocks(runA=ple4, runB=qapply(ple4, function(x) x*1.10))
 plot(pls)
# plot for FLStock, FLPar
 data(ple4)
 rps <- FLPar(Harvest=0.14, Catch=1.29e5, Rec=9.38e5, SSB=1.8e5)
 plot(ple4, rps)
# plot for FLSR
 data(nsher)
 plot(nsher)
}

