% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plot,FLQuant,missing-method}
\alias{plot,FLQuant,missing-method}
\alias{plot,FLQuants,missing-method}
\alias{plot,FLSR,missing-method}
\alias{plot,FLStock,FLPar-method}
\alias{plot,FLStock,missing-method}
\alias{plot,FLStocks,missing-method}
\title{ggplot versions of FLR class plot() methods}
\usage{
\S4method{plot}{FLQuant,missing}(x, main = "", xlab = "", ylab = "",
  na.rm = FALSE, probs = c(0.1, 0.25, 0.5, 0.75, 0.9), type = 7)

\S4method{plot}{FLQuants,missing}(x, main = "", xlab = "", ylab = "",
  probs = c(0.1, 0.25, 0.5, 0.75, 0.9), na.rm = FALSE, type = 7)

\S4method{plot}{FLStock,missing}(x, main = "", xlab = "", ylab = "", ...)

\S4method{plot}{FLStocks,missing}(x, main = "", xlab = "", ylab = "", ...)

\S4method{plot}{FLStock,FLPar}(x, y, ...)

\S4method{plot}{FLSR,missing}(x, y, ...)
}
\arguments{
\item{x}{the coordinates of points in the plot. Alternatively, a
      single plotting structure, function or \emph{any \R object with a
        \code{plot} method} can be provided.}

\item{...}{Arguments to be passed to methods, such as
    \link{graphical parameters} (see \code{\link{par}}).
    Many methods will accept the following arguments:
    \describe{
    \item{\code{type}}{what type of plot should be drawn.  Possible types are
      \itemize{
        \item \code{"p"} for \bold{p}oints,
        \item \code{"l"} for \bold{l}ines,
        \item \code{"b"} for \bold{b}oth,
        \item \code{"c"} for the lines part alone of \code{"b"},
        \item \code{"o"} for both \sQuote{\bold{o}verplotted},
        \item \code{"h"} for \sQuote{\bold{h}istogram} like (or
        \sQuote{high-density}) vertical lines,
        \item \code{"s"} for stair \bold{s}teps,
        \item \code{"S"} for other \bold{s}teps, see \sQuote{Details} below,
        \item \code{"n"} for no plotting.
      }
      All other \code{type}s give a warning or an error; using, e.g.,
      \code{type = "punkte"} being equivalent to \code{type = "p"} for S
      compatibility.  Note that some methods,
      e.g. \code{\link{plot.factor}}, do not accept this.
    }
    \item{\code{main}}{an overall title for the plot: see \code{\link{title}}.}
    \item{\code{sub}}{a sub title for the plot: see \code{\link{title}}.}
    \item{\code{xlab}}{a title for the x axis: see \code{\link{title}}.}
    \item{\code{ylab}}{a title for the y axis: see \code{\link{title}}.}
    \item{\code{asp}}{the \eqn{y/x} aspect ratio,
      see \code{\link{plot.window}}.}
    }
  }

\item{y}{the y coordinates of points in the plot, \emph{optional}
    if \code{x} is an appropriate structure.}
}
\description{
New basic plot methods for some FLR classes are defined in ggplotFL.
}
\examples{
# Plot a single FLQuant
 data(ple4)
 plot(catch.n(ple4))

 # Plot an FLQuant with iters, shows quantiles
 flq <- rnorm(100, catch(ple4), 60000)
 plot(flq)

 # Specify quantiles, default is c(0.10, 0.25, 0.50, 0.75, 0.90)
 plot(flq, probs=c(0.05, 0.40, 0.50, 0.60, 0.95))

 #
# Plot an FLQuants created from ple4 FLStock
 data(ple4)
 plot(FLQuants(SSB=ssb(ple4), rec=rec(ple4)))
# plot of an FLStock
 data(ple4)
 plot(ple4)
# plot for FLStocks
 data(ple4)
 pls <- FLStocks(runA=ple4, runB=qapply(ple4, function(x) x*1.10))
 plot(pls)
# plot for FLStock, FLPar
 data(ple4)
 rps <- FLPar(Harvest=0.14, Catch=1.29e5, Rec=9.38e5, SSB=1.25e6)
 plot(ple4, rps)
# plot for FLSR
 data(nsher)
 plot(nsher)
}

