\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,FLStock,missing-method}
\alias{plot,FLStocks,missing-method}
\title{ggplot versions of FLR class plot() methods}
\description{
  New basic plot for some FLR classes are defined in
  ggplotFL.

  New basic plot for some FLR classes are defined in
  ggplotFL.
}
\examples{
data(ple4)
  plot(ple4)
data(ple4)
  fls <- FLStocks(runA=ple4, runB=ple4)
  plot(fls)
}

